+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/invalid/Cmd.cli.invalid.profile-spec.integration.test.ts
@@ -27,9 +27,9 @@ describe("cmd-cli invalid profile-spec", () => {
 
     it("should fail the command if the profile property is not supplied and the handler requests a profile", () => {
         const response = runCliScript(__dirname + "/__scripts__/profile-spec.sh", TEST_ENVIRONMENT.workingDir);
-        expect(response.status).toBe(1);
         expect(response.stdout.toString()).toBe('');
         expect(response.stderr.toString()).toContain('Internal Error: No profiles of type "blah" were loaded for this command.');
+        expect(response.status).toBe(1);
         expect(response.stderr.toString()).toContain('This error can occur for one of two reasons:');
         expect(response.stderr.toString()).toContain('- The "profile" property on the command definition document ' +
             'does NOT specify the requested profile type');