+++ packages/imperative/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Add client-side custom-event handling capabilities. [#2136](https://github.com/zowe/zowe-cli/pull/2136)
+- Next-Breaking: Refactored the Imperative Event Emitter class. [#2136](https://github.com/zowe/zowe-cli/pull/2136)
+  - Removed the `ImperativeEventEmitter` class.
+  - Added an `EventProcessor` class to handle event listening and emitting.
+  - Added an `EventOperator` class to handle creation and deletion of `EventProcessors`.
+  - Added an `EventUtils` class to contain all common utility methods for the Client Event Handling capabilities.
+  - Added `IEmitter`, `IWatcher`, and `IEmitterAndWatcher` interfaces to expose what application developers should see.
+  
 ## `8.0.0-next.202406201950`
 
 - Enhancement: Added `ProfileInfo.profileManagerWillLoad` function to verify the credential manager can load. [#2111](https://github.com/zowe/zowe-cli/issues/2111)
@@ -71,7 +81,7 @@ All notable changes to the Imperative package will be documented in this file.
 
 ## `8.0.0-next.202405151329`
 
-- Enhancement: Add client-side event handling capabilities. [#1987](https://github.com/zowe/zowe-cli/pull/1987)
+- Enhancement: Add client-side event handling capabilities. [#1987](https://github.com/zowe/zowe-cli/issues/1987)
 
 ## `8.0.0-next.202405061946`